+++ source/core/tests/unit_tests/gold/application/streams/test_calculated_measurements_stream.py
@@ -1,3 +1,4 @@
+import os
 from unittest.mock import Mock
 
 from pyspark.sql import SparkSession
@@ -7,6 +8,7 @@ import core.gold.application.streams.calculated_measurements_stream as sut
 from core.gold.domain.constants.streaming.query_names import QueryNames
 from core.gold.infrastructure.repositories.calculated_measurements_repository import CalculatedMeasurementsRepository
 from core.gold.infrastructure.repositories.gold_measurements_repository import GoldMeasurementsRepository
+from core.gold.infrastructure.repositories.measurements_sap_series_repository import GoldMeasurementsSAPSeriesRepository
 from core.gold.infrastructure.streams.gold_measurements_stream import GoldMeasurementsStream
 
 
@@ -36,6 +38,12 @@ def test__batch_operation__calls_append_to_gold_measurements(mocker: MockerFixtu
     transform_mock = Mock()
     mocker.patch.object(sut, "GoldMeasurementsRepository", return_value=gold_repo_mock)
     mocker.patch.object(sut.transformations, "transform_calculated_to_gold", transform_mock)
+
+    sap_series_repo_mock = Mock(spec=GoldMeasurementsSAPSeriesRepository)
+    transform_sap_mock = Mock()
+    mocker.patch.object(sut, "GoldMeasurementsSAPSeriesRepository", return_value=sap_series_repo_mock)
+    mocker.patch.object(sut.sap_series_transformations, "transform_calculated", transform_sap_mock)
+
     calculated_measurements_mock = Mock()
 
     # Act
@@ -46,3 +54,36 @@ def test__batch_operation__calls_append_to_gold_measurements(mocker: MockerFixtu
     gold_repo_mock.append_if_not_exists.assert_called_once_with(
         transform_mock.return_value, query_name=QueryNames.CALCULATED_TO_GOLD
     )
+
+    transform_sap_mock.assert_called_once_with(calculated_measurements_mock)
+    sap_series_repo_mock.append_if_not_exists.assert_called_once_with(
+        transform_sap_mock.return_value, query_name=QueryNames.CALCULATED_TO_GOLD_SAP_SERIES
+    )
+
+
+def test__batch_operation__when_not_streaming_to_sap_series__should_not_append_to_sap_series(mocker: MockerFixture):
+    # Arrange
+    os.environ["STREAM_CALCULATED_TO_SAP_SERIES"] = "false"
+    gold_repo_mock = Mock(spec=GoldMeasurementsRepository)
+    transform_mock = Mock()
+    mocker.patch.object(sut, "GoldMeasurementsRepository", return_value=gold_repo_mock)
+    mocker.patch.object(sut.transformations, "transform_calculated_to_gold", transform_mock)
+
+    sap_series_repo_mock = Mock(spec=GoldMeasurementsSAPSeriesRepository)
+    transform_sap_mock = Mock()
+    mocker.patch.object(sut, "GoldMeasurementsSAPSeriesRepository", return_value=sap_series_repo_mock)
+    mocker.patch.object(sut.sap_series_transformations, "transform_calculated", transform_sap_mock)
+
+    calculated_measurements_mock = Mock()
+
+    # Act
+    sut._batch_operation(calculated_measurements_mock, 0)
+
+    # Assert
+    transform_mock.assert_called_once_with(calculated_measurements_mock)
+    gold_repo_mock.append_if_not_exists.assert_called_once_with(
+        transform_mock.return_value, query_name=QueryNames.CALCULATED_TO_GOLD
+    )
+
+    transform_sap_mock.assert_not_called()
+    sap_series_repo_mock.assert_not_called()